+++ si-template-projects/si-template-war-project/src/main/java/sipackage/model/TwitterMessages.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stsorg.stsspringframework.stsintegration.model;
+package sipackage.model;
 
 import java.io.Serializable;
 import java.util.ArrayList;
