+++ si-template-projects/si-template-war-project/src/main/resources/META-INF/spring/integration/spring-integration-context.xml
@@ -10,11 +10,11 @@
 		http://www.springframework.org/schema/integration/twitter http://www.springframework.org/schema/integration/twitter/spring-integration-twitter.xsd
 		http://www.springframework.org/schema/context             http://www.springframework.org/schema/context/spring-context.xsd">
 
-	<context:component-scan base-package="stsorg.stsspringframework.stsintegration.service" />
+	<context:component-scan base-package="sipackage" />
 
 	<!-- See also: http://static.springsource.org/spring-integration/reference/htmlsingle/#twitter -->
 	<int-twitter:search-inbound-channel-adapter id="twitter"
-		query="springsource OR springintegration OR cloudfoundry OR grails OR rabbitmq OR #groovy"
+		query="springsource OR springintegration OR cloudfoundry OR grails OR rabbitmq OR groovylang OR springframework"
 		channel="twitterChannel" auto-startup="false">
 		<int:poller fixed-rate="20000" max-messages-per-poll="3" />
 	</int-twitter:search-inbound-channel-adapter>
@@ -25,7 +25,7 @@
 		input-channel="twitterChannel"
 		method="addTwitterMessages" />
 
-	<bean id="twitterService" class="stsorg.stsspringframework.stsintegration.service.impl.DefaultTwitterService" />
+	<bean id="twitterService" class="sipackage.service.impl.DefaultTwitterService" />
 
 	<int:logging-channel-adapter id="logger" level="INFO"
 		expression="'Id:' + payload.id + '; Date:' + payload.createdAt + '; FromUser: ' + payload.fromUser" />
