+++ si-template-projects/si-template-war-project/src/main/resources/log4j.xml
@@ -15,7 +15,7 @@
 		<level value="info" />
 	</logger>
 
-	<logger name="stsorg.stsspringframework.stsintegration">
+	<logger name="sipackage">
 		<level value="info" />
 	</logger>
