+++ si-template-projects/si-template-war-project/src/main/webapp/WEB-INF/wro.properties
