+++ si-template-projects/si-template-war-project/src/test/java/sipackage/JacksonTest.java
@@ -1,4 +1,19 @@
-package stsorg.stsspringframework.stsintegration;
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sipackage;
 
 import java.util.Date;
 
@@ -6,10 +21,16 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import stsorg.stsspringframework.stsintegration.model.TwitterMessage;
-import stsorg.stsspringframework.stsintegration.model.TwitterMessages;
-import stsorg.stsspringframework.stsintegration.support.JaxbJacksonObjectMapper;
+import sipackage.model.TwitterMessage;
+import sipackage.model.TwitterMessages;
+import sipackage.support.JaxbJacksonObjectMapper;
 
+/**
+ *
+ * @author SI-TEMPLATE-AUTHOR
+ * @since  SI-TEMPLATE-VERSION
+ *
+ */
 public class JacksonTest {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(JacksonTest.class);
