+++ si-template-projects/si-template-war-project/wizard.json
@@ -3,23 +3,39 @@
 		"elements" : {
 			"element" :	[
 				{
-					"name" : "mavenGroupId",
-					"description" : "Please enter a 'groupId'",
+					"name" : "SI-TEMPLATE-GROUP-ID",
+ 					"description" : "Please enter a 'groupId'",
 					"type" : "java.lang.String",
 					"page" : 0,
 					"required" : true,
 				},
 				{
-					"name" : "si-template-war-artifactId",
+					"name" : "SI-TEMPLATE-ARTIFACT-ID",
 					"description" : "Please enter an 'artifactId'",
 					"type" : "java.lang.String",
 					"page" : 0,
 					"required" : true,
 				},
 				{
-					"name" : "mavenVersion",
-					"description" : "Please enter a version e.g. '1.0.0.BUILD-SNAPSHOT'",
+					"name" : "SI-TEMPLATE-BUILD-VERSION",
+					"description" : "Please enter a build version e.g. '1.0.0.BUILD-SNAPSHOT'",
 					"defaultValue": "1.0.0.BUILD-SNAPSHOT",
+					"type" : "java.lang.String",
+					"page" : 0,
+					"required" : true,
+				},
+				{
+					"name" : "SI-TEMPLATE-VERSION",
+					"description" : "Please enter a short version e.g. '1.0' (For JavaDocs)",
+					"defaultValue": "1.0",
+					"type" : "java.lang.String",
+					"page" : 0,
+					"required" : true,
+				},
+				{
+					"name" : "SI-TEMPLATE-AUTHOR",
+					"description" : "Please enter the author name",
+					"defaultValue": "Your name",
 					"type" : "java.lang.String",
 					"page" : 0,
 					"required" : true,
@@ -30,13 +46,13 @@
 					"type" : "java.lang.String",
 					"page" : 0,
 					"required" : true,
-					"pattern" : "\\w+\\.\\w+(\\.\\w+)+",
+					"pattern" : "\\w+(\\.\\w+)*",
 					"replaceKind" : "topLevelPackage"
 				}
 			]
 		},
-		"topLevelPackage" : "stsorg.stsspringframework.stsintegration.mysubsystem",
-		"projectName" : "si-template-project-war",
+		"topLevelPackage" : "sipackage",
+		"projectName" : "si-template-project-name",
 		"pages" : {
 			"page" : [
 				{
